k8s: Refactor apis/cilium.io/utils to apis/utils
As we create more k8s API groups it would be nice to reuse common
functions like the CRD update. No functionality change should occur.
